refactor: GRAPHIC mode 6,7 now assume convention addressing byte order
The SDRAM module exposes a 16bit data bus. As the VDP uses an 8 bit data model,
the 16bits is split into 2 bytes. Where, prior to this commit, the state of A16 address bit
would be use to select either the high or low byte from the single addressed word.

So in the first stored word, we would have the byte for Address 0x00000 and the byte for address 0x10000.
Then the next word would contain the byte for Address 0x00001 and 0x10001

My perferred model is to make it more conventional, so that:

First word would store data for Address 0x00000 and 0x00001, and the next word would store 0x00002 and 0x00003 and so on.

The previously implemenation in Graphics mode 6 and 7 address translation would 'undo' this addressing schema and apply a
convention order.  It did this by swapping A16 and A0.

So this commit removes the double operation.  Memory is addressed in a convention linear fasion, and the Graphics 6 and 7
no longer need to undo this addressing order.

(Graphics 6 and 7 seems to use the double byte retrieval for its YJK coloring decoding.  This is where 2x2 pixels have seperate
chroma component and a single colour component.)